Move HealthCheckInterface from ::grpc_impl to ::grpc
Revert #18371
