+++ tools/doxygen/Doxyfile.c++.internal
@@ -948,7 +948,6 @@ include/grpcpp/generic/generic_stub.h \
 include/grpcpp/generic/generic_stub_impl.h \
 include/grpcpp/grpcpp.h \
 include/grpcpp/health_check_service_interface.h \
-include/grpcpp/health_check_service_interface_impl.h \
 include/grpcpp/impl/call.h \
 include/grpcpp/impl/channel_argument_option.h \
 include/grpcpp/impl/client_unary_call.h \
